Update README.md (#168)
Stackbuilder is probably not sufficient, to say because ogr_fdw isn't a separate install.  I bundle it with PostGIS installer since it relies on GDAL and I can only have one GDAL install. (so I added a note it's part of PostGIS Bundle) so people don't get confused looking for it separately.